+++ src/IdealGraphics/IdealGraphics.vcxproj.filters
@@ -140,8 +140,8 @@
     <ClInclude Include="GraphicsEngine\public\IDirectionalLight.h" />
     <ClInclude Include="GraphicsEngine\public\ILight.h" />
     <ClInclude Include="GraphicsEngine\public\IdealRendererDefinitions.h" />
-    <ClInclude Include="Editor\imgui\imconfig.h" />
-    <ClInclude Include="Editor\imgui\imgui.h" />
+    <ClInclude Include="GraphicsEngine\public\imconfig.h" />
+    <ClInclude Include="GraphicsEngine\public\imgui.h" />
     <ClInclude Include="Editor\imgui\imgui_impl_dx12.h" />
     <ClInclude Include="Editor\imgui\imgui_impl_win32.h" />
     <ClInclude Include="Editor\imgui\imgui_internal.h" />
